+++ src/app/(whitebg)/users/_components/join/JoinAgreement.tsx
