+++ src/app/(whitebg)/users/_components/join/JoinReceiveNotification.tsx
